renew code

+++ python/fleetx/utils/downloader.py
@@ -45,7 +45,8 @@ def untar_files_with_check(local_path, trainer_id, trainer_num,
                     if os.path.exists("{}/{}".format(local_path, ff)):
                         ready_filelist.append("{}/{}".format(local_path, ff))
                 if len(ready_filelist) == len(filelist):
-                    num_per_trainer = (len(ready_filelist) / trainer_num) + 1
+                    num_per_trainer = int(len(ready_filelist) /
+                                          trainer_num) + 1
                     if (trainer_id + 1
                         ) * num_per_trainer < len(ready_filelist):
                         sub_list = ready_filelist[trainer_id * num_per_trainer:
